requirements.txt
streamlit
sentence-transformers
shap
numpy
python-dotenv
groq
torch
transformers
pdfplumber
python-docx
